refactor: reorganiza arquitetura para MVVM + Clean Code com separação aluno/adm
- Cria estrutura feature/aluno/ e feature/adm/ separando as áreas do sistema
- Move todas as telas do aluno para feature/aluno/{acervo,estante,eventos,home,livro,notificacao,perfil}/ui/
- Extrai CapaLivro para core/components/ (componente global reutilizável)
- Extrai FiltroAvancado para core/components/ (bottom sheet global)
- Move LivrosSalvosState e catalogoGlobal para core/data/ com package correto
- Centraliza paleta de cores (AzulPrimario, CinzaTexto, coresFallback) em core/designsystem/Color.kt
- Corrige todos os imports e package declarations em cascata
- Remove pastas duplicadas/com typo (feature/aluloo/, feature/aluno/ui/ plana)
- Atualiza ForLibraryApp.kt com os novos caminhos de import
- Atualiza README com descrição completa da arquitetura e de cada pasta
