+++ app/src/main/java/com/br/unifor/for_library/core/components/FiltroAvancado.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Close
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
@@ -17,7 +16,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.br.unifor.for_library.feature.home.ui.AzulPrimario
+import com.br.unifor.for_library.core.designsystem.AzulPrimario
 
 enum class OrdemFiltro(val label: String) {
     MAIS_RECENTES("Mais Recentes"),
